fix: values may be floats
This fixes the following deprecation notices:

Implicit conversion from float to int loses precision in TtfParser.php:575
Implicit conversion from float to int loses precision in TtfParser.php:576

+++ src/Pdf/TtfParser.php
@@ -106,8 +106,8 @@ class TtfParser
     public int $stemV;
     public int $italicAngle;
     public int $flags;
-    public int $underlinePosition;
-    public int $underlineThickness;
+    public float $underlinePosition;
+    public float $underlineThickness;
     public string $charWidths;
     public float $defaultWidth;
     public int $maxStrLenRead;
